fix: moving functions to a separate file
